refactor: move validate from registration to helpers file
